# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Resilium API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me/plans:
    get:
      tags: [users]
      operationId: getMyPlans
      summary: Get the authenticated user's saved plans
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: List of saved plans
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyPlansResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /users/me/plans/{reportId}:
    delete:
      tags: [users]
      operationId: deleteMyPlan
      summary: Delete a saved plan belonging to the authenticated user
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Plan deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Plan not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  parameters:
...
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    PlanSummary:
      type: object
      required: [reportId, createdAt, scoreOverall]
      properties:
        reportId:
          type: string
        createdAt:
          type: string
          format: date-time
        scoreOverall:
          type: number
    MyPlansResponse:
      type: object
      required: [plans]
      properties:
        plans:
          type: array
          items:
            $ref: '#/components/schemas/PlanSummary'
...